Added OntologyLinkInstance, which extends SchemaInstance.
OntologyLinkInstance not only tracks "content" (XML tags), but also
"value" (actual value used by AttributeInstance). This is done so that
when an attribute is promoted to be an ontology schema, but later
demoted back to be an attribute due to one-to-one relation with parent
schema (schema flattening), the "value" of the attribute instance is not
lost.